requirements.txt
dnspython
tldextract
browser
